requirements.txt
pure-python-adb==0.3.0.dev0
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0